refactor(store): extract file-store commands into StoreCommands mixin
Move the fs_* file-storage DAP commands out of TaskCommands (cmd_task.py)
into a dedicated StoreCommands mixin (cmd_store.py), mirroring the module
layout already present on feat/alb.

Why:
  feat/alb relocated these same fs_* handlers from cmd_task.py into a new
  cmd_store.py as part of its larger pod/RequestContext refactor. Because
  feat/fileexplorer kept them in cmd_task.py and extended them (Explorer's
  download support), every rebase between the two branches produced a large
  add-block/delete-block conflict over the entire fs_* section. Aligning the
  file layout here shrinks that recurring conflict to just the ctx-parameter
  lines, which are small and uniform.

What changed:
  - cmd_store.py (new): StoreCommands(DAPConn) owning on_rrext_store, the
    _store_subcommand_handlers dispatch table, _get_file_store, and all
    _store_fs_* handlers. This is feat/alb's StoreCommands with the alb-only
    RequestContext threading removed, since feat/fileexplorer's dispatch layer
    calls handlers with (request) only:
      * verify_permission('task.store', ctx) -> verify_permission('task.store')
      * _get_file_store(ctx) -> _get_file_store() scoped via self._account_info.userId
      * dropped the alb-only RequestContext/DAPRequest/DAPResponse imports
  - _store_fs_geturl retains Explorer's download_name passthrough (alb's copy
    predates it), so get_url still forwards download_name for the
    Content-Disposition: attachment path. Verified end to end (client SDKs ->
    dispatch -> file_store.get_url -> S3/Azure content_disposition and local
    /task/fetch header).
  - cmd_task.py: removed the moved on_rrext_store/_get_file_store/_store_fs_*
    code and the dispatch-table population in __init__ (now `pass`). No
    imports orphaned.
  - task_conn.py: wired StoreCommands into TaskConn at feat/alb's positions
    (import, base list, and the required StoreCommands.__init__ call that
    builds the dispatch table).
  - test_cmd_task.py: repointed the moved symbols from TaskCommands to
    StoreCommands, bound _get_file_store onto the __init__-bypassed stub via
    MethodType, and moved the constructor test to StoreCommands.__init__.
    17/17 tests pass.
