+++ packages/ai/src/ai/modules/task/commands/cmd_task.py
@@ -104,22 +104,7 @@ class TaskCommands(DAPConn):
             transport (TransportBase): Communication transport layer for DAP messages
             **kwargs: Additional arguments passed to parent DAPConn constructor
         """
-        # Map of store subcommand names to handler methods.
-        # Populated here so new subcommands can be added without touching the
-        # dispatcher logic in on_rrext_store.
-        self._store_subcommand_handlers = {
-            'fs_open': self._store_fs_open,
-            'fs_read': self._store_fs_read,
-            'fs_write': self._store_fs_write,
-            'fs_close': self._store_fs_close,
-            'fs_delete': self._store_fs_delete,
-            'fs_list_dir': self._store_fs_list_dir,
-            'fs_mkdir': self._store_fs_mkdir,
-            'fs_rmdir': self._store_fs_rmdir,
-            'fs_stat': self._store_fs_stat,
-            'fs_rename': self._store_fs_rename,
-            'fs_geturl': self._store_fs_geturl,
-        }
+        pass
 
     async def on_execute(self, request: Dict[str, Any]) -> Dict[str, Any]:
         """
@@ -434,337 +419,3 @@ class TaskCommands(DAPConn):
             # Log and re-raise for standard error handling
             self.debug_message(f'Failed to list tasks: {str(e)}')
             raise
-
-    async def on_rrext_store(self, request: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Handle DAP 'rrext_store' command - unified project and template storage operations.
-
-        Central dispatcher for all storage subcommands. Verifies permissions once
-        and routes to appropriate subcommand handler.
-
-        Args:
-            request (Dict[str, Any]): DAP request containing:
-                - arguments: Dict with:
-                    - subcommand: One of 'save_project', 'get_project', 'delete_project', 'get_all_projects',
-                                  'save_template', 'get_template', 'delete_template', 'get_all_templates'
-                    - ...: Additional arguments specific to subcommand
-
-        Returns:
-            Dict[str, Any]: DAP response (format depends on subcommand)
-
-        Raises:
-            ValueError: If subcommand is missing or unknown
-            Exception: If subcommand execution fails
-        """
-        try:
-            # Require store permission (once for all subcommands)
-            self.verify_permission('task.store')
-
-            # Extract subcommand
-            args = request.get('arguments', {})
-            subcommand = args.get('subcommand')
-
-            if not subcommand:
-                raise ValueError('Subcommand is required')
-
-            # Dispatch to appropriate handler using the pre-built lookup dict.
-            # The walrus operator assigns the handler if found; None triggers else.
-            if handler := self._store_subcommand_handlers.get(subcommand):
-                return await handler(request, args)
-            else:
-                raise ValueError(f'Unknown subcommand: {subcommand}')
-
-        except Exception as e:
-            self.debug_message(f'Store operation failed: {str(e)}')
-            raise
-
-    def _get_file_store(self):
-        """
-        Get a FileStore scoped to the authenticated user.
-
-        Returns:
-            FileStore: A file-store instance that isolates all paths under
-                the current user's storage namespace.
-        """
-        # Scope the file store to the calling user so users cannot access each
-        # other's files through the store API.
-        return self._server.store.get_file_store(self._account_info.userId)
-
-    # =========================================================================
-    # Generic File Store Handlers
-    # =========================================================================
-
-    async def _store_fs_open(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Open a file handle for reading or writing.
-
-        For write mode (``mode='w'``) the backend creates a new write handle and
-        returns its ID.  For read mode (default) the backend opens the file,
-        validates it exists, and returns metadata alongside the handle ID.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``path``.  Optional ``mode``
-                (``'r'`` or ``'w'``, defaults to ``'r'``).
-
-        Returns:
-            Dict[str, Any]: DAP response.
-                Write mode: ``body.handle`` — the new write handle ID.
-                Read mode:  result from ``fs.open_read`` (includes handle + metadata).
-        """
-        fs = self._get_file_store()
-        path = args.get('path')
-        mode = args.get('mode', 'r')
-
-        if mode == 'w':
-            # Create a write handle; the connection_id is used to tie the handle
-            # lifetime to this connection so it is cleaned up on disconnect.
-            handle_id = await fs.open_write(path, self._connection_id)
-            return self.build_response(request, body={'handle': handle_id})
-        else:
-            # Open for reading; returns handle ID plus file metadata (size, etc.).
-            result = await fs.open_read(path, self._connection_id)
-            return self.build_response(request, body=result)
-
-    async def _store_fs_read(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Read data from an open read handle.
-
-        Clamps ``offset`` and ``length`` to safe values before forwarding to the
-        backend so that a misbehaving client cannot request an unbounded read.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``handle``.  Optional ``offset``
-                (default 0) and ``length`` (default and max 4 MiB).
-
-        Returns:
-            Dict[str, Any]: DAP response with ``body.size`` (bytes read) and
-                ``arguments.data`` (the raw bytes).
-        """
-        fs = self._get_file_store()
-        handle = args.get('handle')
-        offset = args.get('offset', 0)
-        length = args.get('length', 4_194_304)
-
-        # Clamp client-supplied values to safe defaults.
-        # Negative or non-integer offsets are reset to 0.
-        if not isinstance(offset, int) or offset < 0:
-            offset = 0
-        # Non-positive or non-integer lengths are reset to the default chunk size.
-        if not isinstance(length, int) or length <= 0:
-            length = 4_194_304
-        # Cap the length at 4 MiB to prevent memory exhaustion from large reads.
-        length = min(length, 4_194_304)
-
-        # Read the chunk from the file store backend.
-        data = await fs.read_chunk(handle, offset, length, connection_id=self._connection_id)
-
-        # Build the response: body carries the byte count; arguments carries the
-        # raw data separately so the protocol can handle binary payloads correctly.
-        response = self.build_response(request, body={'size': len(data)})
-        response['arguments'] = {'data': data}
-        return response
-
-    async def _store_fs_write(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Write data to an open write handle.
-
-        Accepts both ``bytes`` and ``str`` data; strings are UTF-8 encoded before
-        being forwarded to the backend.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``handle``.  Optional ``data``
-                (bytes or str, defaults to empty bytes).
-
-        Returns:
-            Dict[str, Any]: DAP response with ``body.bytesWritten`` containing the
-                number of bytes actually written.
-        """
-        fs = self._get_file_store()
-        handle = args.get('handle')
-        data = args.get('data', b'')
-
-        # Normalise string data to bytes so the backend always receives bytes.
-        if isinstance(data, str):
-            data = data.encode('utf-8')
-
-        # Write the chunk and return the actual byte count confirmed by the backend.
-        written = await fs.write_chunk(handle, data, connection_id=self._connection_id)
-        return self.build_response(request, body={'bytesWritten': written})
-
-    async def _store_fs_close(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Close a file handle.
-
-        Dispatches to the correct close method based on the handle mode so that
-        the backend can flush write buffers for write handles.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``handle``.  Optional ``mode``
-                (``'r'`` or ``'w'``, defaults to ``'r'``).
-
-        Returns:
-            Dict[str, Any]: Empty success DAP response.
-        """
-        fs = self._get_file_store()
-        handle = args.get('handle')
-        mode = args.get('mode', 'r')
-
-        if mode == 'w':
-            # Flush and close a write handle; triggers any finalisation (e.g. S3 upload).
-            await fs.close_write(handle, connection_id=self._connection_id)
-        else:
-            # Release a read handle and free associated resources.
-            await fs.close_read(handle, connection_id=self._connection_id)
-        return self.build_response(request)
-
-    async def _store_fs_delete(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Delete file from file store.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``path`` — the file to delete.
-
-        Returns:
-            Dict[str, Any]: Empty success DAP response.
-        """
-        # Delegate deletion to the user-scoped file store.
-        await self._get_file_store().delete(args.get('path'))
-        return self.build_response(request)
-
-    async def _store_fs_list_dir(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        List directory contents.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Optional ``path`` — directory to list
-                (defaults to the root of the user's store, i.e. ``''``).
-
-        Returns:
-            Dict[str, Any]: DAP response with directory listing as the body.
-        """
-        # Delegate to the file store; default to the root path if not specified.
-        result = await self._get_file_store().list_dir(args.get('path', ''))
-        return self.build_response(request, body=result)
-
-    async def _store_fs_mkdir(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Create directory.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``path`` — the directory to create.
-
-        Returns:
-            Dict[str, Any]: Empty success DAP response.
-        """
-        # Create the directory in the user-scoped file store.
-        await self._get_file_store().mkdir(args.get('path'))
-        return self.build_response(request)
-
-    async def _store_fs_rmdir(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Remove directory.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``path`` (non-empty string). Optional
-                ``recursive`` (strict bool; any non-bool value is rejected rather than
-                coerced). If ``True`` the directory and all its contents are removed;
-                if ``False`` the call fails when the directory is non-empty.
-
-        Returns:
-            Dict[str, Any]: Empty success DAP response, or a DAP error response if
-                ``path`` is missing/empty or ``recursive`` is not a bool.
-        """
-        # Validate path is a non-empty string before touching the store — the
-        # FileStore layer rejects empty paths, but an early error here produces
-        # a cleaner DAP-level message.
-        path = args.get('path')
-        if not isinstance(path, str) or not path:
-            return self.build_error(request, 'rmdir requires a non-empty "path" string')
-
-        # Accept only an explicit bool for ``recursive`` — avoid silently enabling
-        # recursive deletion when the client sends a non-bool truthy value (e.g.
-        # a string or a dict).
-        recursive = args.get('recursive', False)
-        if not isinstance(recursive, bool):
-            return self.build_error(request, 'rmdir "recursive" must be a boolean')
-
-        # Remove the directory; pass the recursive flag from the client.
-        await self._get_file_store().rmdir(path, recursive=recursive)
-        return self.build_response(request)
-
-    async def _store_fs_stat(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Get file/directory metadata.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``path`` — the path to stat.
-
-        Returns:
-            Dict[str, Any]: DAP response with metadata (size, modified time, type,
-                etc.) as the body.
-        """
-        # Retrieve metadata for the given path from the user-scoped store.
-        result = await self._get_file_store().stat(args.get('path'))
-        return self.build_response(request, body=result)
-
-    async def _store_fs_rename(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Rename a file or directory.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``old_path`` (current path) and
-                ``new_path`` (desired path). Both are required non-empty strings.
-
-        Returns:
-            Dict[str, Any]: Empty success DAP response, or a DAP error response if
-                either argument is missing or not a non-empty string.
-        """
-        # Validate both paths up front so FileStore.rename does not receive None
-        # and fail with a less actionable error further down the stack.
-        old_path = args.get('old_path')
-        new_path = args.get('new_path')
-        if not isinstance(old_path, str) or not old_path:
-            return self.build_error(request, 'rename requires a non-empty "old_path" string')
-        if not isinstance(new_path, str) or not new_path:
-            return self.build_error(request, 'rename requires a non-empty "new_path" string')
-
-        # Delegate the rename operation to the user-scoped file store backend.
-        await self._get_file_store().rename(old_path, new_path)
-        return self.build_response(request)
-
-    async def _store_fs_geturl(self, request: Dict[str, Any], args: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Get a direct HTTP URL for accessing a file in the store.
-
-        For cloud backends (S3, Azure), returns a presigned/SAS URL.
-        For filesystem backends, returns a JWT-signed ``/task/fetch`` URL
-        that the server's HTTP endpoint validates and serves.
-
-        Args:
-            request (Dict[str, Any]): Original DAP request.
-            args (Dict[str, Any]): Must contain ``path`` (relative store path).
-                Optional ``expires_in`` (seconds, default 3600) and
-                ``download_name`` (forces a browser download with this filename
-                via ``Content-Disposition: attachment``).
-
-        Returns:
-            Dict[str, Any]: DAP response with ``url`` in the body.
-        """
-        path = args.get('path')
-        if not isinstance(path, str) or not path:
-            return self.build_error(request, 'geturl requires a non-empty "path" string')
-
-        expires_in = int(args.get('expires_in', 3600))
-        download_name = args.get('download_name') or None
-        url = await self._get_file_store().get_url(path, expires_in, download_name=download_name)
-        return self.build_response(request, body={'url': url})
