+++ packages/ai/src/ai/modules/task/task_conn.py
@@ -74,6 +74,7 @@ from .commands.cmd_account import AccountCommands
 from .commands.cmd_app import AppCommands
 from .commands.cmd_public import PublicCommands
 from .commands.cmd_deploy import DeployCommands
+from .commands.cmd_store import StoreCommands
 from ai.account.models import AccountInfo, resolve_task_permissions, resolve_team_permissions
 from ai.common.account import AccountPipelineValidation
 
@@ -103,6 +104,7 @@ class TaskConn(
     AppCommands,
     PublicCommands,
     DeployCommands,
+    StoreCommands,
     DAPConn,
 ):
     """
@@ -191,6 +193,7 @@ class TaskConn(
         CProfileCommands.__init__(self, connection_id, server, transport, **kwargs)
         AccountCommands.__init__(self, connection_id, server, transport, **kwargs)
         AppCommands.__init__(self, connection_id, server, transport, **kwargs)
+        StoreCommands.__init__(self, connection_id, server, transport, **kwargs)
 
         # Store connection identifier for tracking and logging
         self._connection_id = connection_id
